+++ dwds/CHANGELOG.md
@@ -1,4 +1,6 @@
 ## 14.0.1-dev
+- Add `libraryFilters` optional parameter to the vm service implememtation
+  of `getSourceReport`.
 - Update temp variable detection heuristics so internal JS type objects do
   not show in the debugger.